+++ packages/article-comments/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.17](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.6.16...@times-components/article-comments@0.6.17) (2020-02-13)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
 ## [0.6.16](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.6.15...@times-components/article-comments@0.6.16) (2020-02-07)
 
 **Note:** Version bump only for package @times-components/article-comments