+++ packages/topic/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.1.35](https://github.com/newsuk/times-components/compare/@times-components/topic@5.1.34...@times-components/topic@5.1.35) (2020-02-13)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
 ## [5.1.34](https://github.com/newsuk/times-components/compare/@times-components/topic@5.1.33...@times-components/topic@5.1.34) (2020-02-10)
 
 **Note:** Version bump only for package @times-components/topic